divided organisms into seperate files
