+++ src/animal.js
@@ -1,0 +1,12 @@
+import {Organism} from './organism.js';
+
+export class Animal extends Organism {
+    constructor() {
+        super();
+        this.strenght = 0;
+    }
+
+    mate() {}
+
+    fight() {}
+}
