+++ src/fox.js
@@ -1,0 +1,10 @@
+import {Animal} from './animal.js';
+
+export class Fox extends Animal {
+    constructor() {
+        super();
+        this.initiative = 7;
+        this.strenght = 4;
+        this.name = "fox";
+    }
+}
