+++ src/grass.js
@@ -1,0 +1,8 @@
+import {Plant} from './plant.js';
+
+export class Grass extends Plant {
+    constructor() {
+        super();
+        this.name = "grass";
+    }
+}
