+++ src/guarana.js
@@ -1,0 +1,8 @@
+import {Plant} from './plant.js';
+
+export class Guarana extends Plant {
+    constructor() {
+        super();
+        this.name = "guarana";
+    }
+}
