+++ src/organismsList.js
@@ -1,0 +1,12 @@
+import {Sheep} from "./sheep";
+import {Wolf} from "./wolf";
+import {Fox} from "./fox";
+import {Antelope} from "./antelope";
+import {Turtle} from "./turtle";
+import {Grass} from "./grass";
+import {Guarana} from "./guarana";
+import {PoisonBerry} from "./poisonBerry";
+import {SowThistle} from "./snowThistle";
+
+
+export let organismsList = [Wolf, Sheep, Fox, Antelope, Turtle, Grass, Guarana, PoisonBerry, SowThistle];
