+++ src/plant.js
@@ -1,0 +1,10 @@
+import {Organism} from './organism.js';
+
+export class Plant extends Organism {
+    constructor() {
+        super();
+        this.name = "plant";
+    }
+
+    spread() {}
+}
