+++ src/poisonBerry.js
@@ -1,0 +1,8 @@
+import {Plant} from './plant.js';
+
+export class PoisonBerry extends Plant {
+    constructor() {
+        super();
+        this.name = "poison berry";
+    }
+}
