+++ src/sheep.js
@@ -1,0 +1,10 @@
+import {Animal} from './animal.js';
+
+export class Sheep extends Animal {
+    constructor() {
+        super();
+        this.initiative = 4;
+        this.strenght = 3;
+        this.name = "sheep";
+    }
+}
