+++ src/snowThistle.js
@@ -1,0 +1,8 @@
+import {Plant} from './plant.js';
+
+export class SowThistle extends Plant {
+    constructor() {
+        super();
+        this.name = "snow thistle";
+    }
+}
