+++ src/turtle.js
@@ -1,0 +1,10 @@
+import {Animal} from './animal.js';
+
+export class Turtle extends Animal {
+    constructor() {
+        super();
+        this.initiative = 1;
+        this.strenght = 2;
+        this.name = "turtle";
+    }
+}
