+++ src/wolf.js
@@ -1,0 +1,10 @@
+import {Animal} from './animal.js';
+
+export class Wolf extends Animal {
+    constructor() {
+        super();
+        this.initiative = 5;
+        this.strenght = 9;
+        this.name = "wolf";
+    }
+}
